refactor(liblili2json): move get_skills
