+++ workspaces/liblili2json/liblili2json/__init__.py
@@ -23,16 +23,3 @@ def get_challenges_data(skill, course):
         make_challenges_using(get_phrase_challenges, skill.phrases, course),
         make_challenges_using(get_word_challenges, skill.words, course),
     ], start=[])
-
-
-def get_skill_data(skill, course):
-    """
-    Format Course according to the JSON structure
-    """
-
-    return {
-        "id": get_opaque_id(skill, "Skill"),
-        "levels": calculate_number_of_levels(
-            len(skill.words), len(skill.phrases)),
-        "challenges": get_challenges_data(skill, course),
-    }
