+++ workspaces/liblili2json/liblili2json/skills.py
@@ -1,0 +1,15 @@
+from .utils import get_opaque_id, calculate_number_of_levels
+from . import get_challenges_data
+
+
+def get_skill_data(skill, course):
+    """
+    Format Course according to the JSON structure
+    """
+
+    return {
+        "id": get_opaque_id(skill, "Skill"),
+        "levels": calculate_number_of_levels(
+            len(skill.words), len(skill.phrases)),
+        "challenges": get_challenges_data(skill, course),
+    }
