+++ workspaces/liblili2json/tests/test_liblili2json.py
@@ -1,7 +1,7 @@
 from unittest.mock import patch
 from unittest import TestCase
 from liblili2json.course import get_course_data
-from liblili2json import get_skill_data
+from liblili2json.skills import get_skill_data
 from liblili2json import get_challenges_data
 from liblili2json.utils import calculate_number_of_levels
 from liblili2json import get_word_challenges
@@ -131,26 +131,26 @@ class TestGetSkillData(TestCase):
             "challenges": []
         }
 
-    @patch('liblili2json.calculate_number_of_levels')
+    @patch('liblili2json.skills.calculate_number_of_levels')
     def test_correct_number_of_levels(self, mock):
         FAKE_NUMBER = "fake number"
         mock.return_value = FAKE_NUMBER
         converted_skill = get_skill_data(fakes.emptySkill, fakes.course1)
         assert converted_skill["levels"] == FAKE_NUMBER
 
-    @patch('liblili2json.calculate_number_of_levels')
+    @patch('liblili2json.skills.calculate_number_of_levels')
     def test_calculates_levels_correctly(self, mock):
         get_skill_data(fakes.skills[1], fakes.course1)
         mock.assert_called_with(4, 1)
 
-    @patch('liblili2json.get_challenges_data')
+    @patch('liblili2json.skills.get_challenges_data')
     def test_correct_challenges(self, mock):
         FAKE_CHALLENGES = "fake challenges"
         mock.return_value = FAKE_CHALLENGES
         converted_skill = get_skill_data(fakes.skills[1], fakes.course1)
         assert converted_skill["challenges"] == FAKE_CHALLENGES
 
-    @patch('liblili2json.get_challenges_data')
+    @patch('liblili2json.skills.get_challenges_data')
     def test_formats_challenges_correctly(self, mock):
         get_skill_data(fakes.skills[1], fakes.course1)
         mock.assert_called_with(fakes.skills[1], fakes.course1)
